Move ContractDeployer to a separate file
